pyproject.toml
[project]
name = "pleasanter-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
